backend/requirements.txt
fastapi
uvicorn
groq
pyyaml
python-dotenv
